pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "talabat-driver-wallet"
version = "1.0.0"
description = "Professional TUI accounting system for delivery drivers"
authors = [{name = "Talabat Driver"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.34.0",
    "rich>=13.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "arabic-reshaper>=3.0.0",
    "python-bidi>=0.4.2",
]

[project.scripts]
work = "talabat_wallet.__main__:main"